[PC-253] 패키지 재정리, 화면 생성
