+++ feature/auth/src/main/java/com/puzzle/auth/navigation/AuthNavigation.kt
@@ -3,7 +3,7 @@ package com.puzzle.auth.navigation
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
-import com.puzzle.auth.AuthRoute
+import com.puzzle.auth.page.main.AuthRoute
 import com.puzzle.navigation.AuthGraph
 import com.puzzle.navigation.AuthGraphDest
