+++ feature/auth/src/main/java/com/puzzle/auth/page/main/contract/AuthIntent.kt
@@ -1,0 +1,4 @@
+package com.puzzle.auth.page.main.contract
+
+class AuthIntent {
+}
