+++ feature/auth/src/main/java/com/puzzle/auth/page/registration/contract/AuthRegistrationIntent.kt
@@ -1,0 +1,4 @@
+package com.puzzle.auth.page.registration.contract
+
+sealed class AuthRegistrationIntent {
+}
