+++ feature/auth/src/main/java/com/puzzle/auth/page/registration/contract/AuthRegistrationSideEffect.kt
@@ -1,0 +1,4 @@
+package com.puzzle.auth.page.registration.contract
+
+sealed class AuthRegistrationSideEffect {
+}
