+++ feature/auth/src/main/java/com/puzzle/auth/page/registration/contract/AuthRegistrationState.kt
@@ -1,0 +1,7 @@
+package com.puzzle.auth.page.registration.contract
+
+import com.airbnb.mvrx.MavericksState
+
+data class AuthRegistrationState(
+    val a: Boolean,
+) : MavericksState
