+++ feature/auth/src/main/java/com/puzzle/auth/page/verification/contract/AuthVerificationIntent.kt
@@ -1,0 +1,4 @@
+package com.puzzle.auth.page.verification.contract
+
+sealed class AuthVerificationIntent {
+}
