+++ feature/matching/src/main/java/com/puzzle/matching/common/constant/DialogType.kt
@@ -1,0 +1,7 @@
+package com.puzzle.matching.common.constant
+
+enum class DialogType {
+    ACCEPT_MATCHING,
+    DECLINE_MATCHING,
+    PROFILE_IMAGE_DETAIL,
+}
