--- feature/matching/src/main/java/com/puzzle/matching/contract/MatchingIntent.kt
@@ -1,5 +1,0 @@
-package com.puzzle.matching.contract
-
-sealed class MatchingIntent {
-    data object NavigateToMatchingDetail : MatchingIntent()
-}
