--- feature/matching/src/main/java/com/puzzle/matching/contract/MatchingSideEffect.kt
@@ -1,3 +1,0 @@
-package com.puzzle.matching.contract
-
-sealed class MatchingSideEffect
