+++ feature/matching/src/main/java/com/puzzle/matching/page/detail/MatchingDetailViewModel.kt
@@ -1,0 +1,81 @@
+package com.puzzle.matching.page.detail
+
+import com.airbnb.mvrx.MavericksViewModel
+import com.airbnb.mvrx.MavericksViewModelFactory
+import com.airbnb.mvrx.hilt.AssistedViewModelFactory
+import com.airbnb.mvrx.hilt.hiltMavericksViewModelFactory
+import com.puzzle.matching.page.detail.contract.MatchingDetailIntent
+import com.puzzle.matching.page.detail.contract.MatchingDetailSideEffect
+import com.puzzle.matching.page.detail.contract.MatchingDetailState
+import com.puzzle.navigation.NavigationEvent
+import com.puzzle.navigation.NavigationHelper
+import dagger.assisted.Assisted
+import dagger.assisted.AssistedFactory
+import dagger.assisted.AssistedInject
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.Channel.Factory.BUFFERED
+import kotlinx.coroutines.flow.launchIn
+import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.flow.receiveAsFlow
+import kotlinx.coroutines.launch
+
+class MatchingDetailViewModel @AssistedInject constructor(
+    @Assisted initialState: MatchingDetailState,
+    private val navigationHelper: NavigationHelper,
+) : MavericksViewModel<MatchingDetailState>(initialState) {
+
+    private val intents = Channel<MatchingDetailIntent>(BUFFERED)
+
+    private val _sideEffect = Channel<MatchingDetailSideEffect>(BUFFERED)
+    val sideEffect = _sideEffect.receiveAsFlow()
+
+    init {
+        intents.receiveAsFlow()
+            .onEach(::processIntent)
+            .launchIn(viewModelScope)
+    }
+
+    internal fun onIntent(intent: MatchingDetailIntent) = viewModelScope.launch {
+        intents.send(intent)
+    }
+
+    private fun processIntent(intent: MatchingDetailIntent) {
+        when (intent) {
+            MatchingDetailIntent.OnMatchingDetailCloseClick -> processOnMatchingDetailCloseClickIntent()
+            MatchingDetailIntent.OnPreviousPageClick -> processOnPreviousPageClickIntent()
+            MatchingDetailIntent.OnNextPageClick -> processOnNextPageClickIntent()
+            MatchingDetailIntent.OnMoreClick -> Unit
+        }
+    }
+
+    private fun processOnNextPageClickIntent() {
+        setState {
+            copy(currentPage = MatchingDetailState.MatchingDetailPage.getNextPage(currentPage))
+        }
+    }
+
+    private fun processOnPreviousPageClickIntent() {
+        setState {
+            copy(currentPage = MatchingDetailState.MatchingDetailPage.getPreviousPage(currentPage))
+        }
+    }
+
+    private fun processOnMatchingDetailCloseClickIntent() {
+        navigationHelper.navigate(NavigationEvent.NavigateUp)
+    }
+
+    private fun handleSideEffect(sideEffect: MatchingDetailSideEffect) {
+        when (sideEffect) {
+            else -> Unit
+        }
+    }
+
+    @AssistedFactory
+    interface Factory : AssistedViewModelFactory<MatchingDetailViewModel, MatchingDetailState> {
+        override fun create(state: MatchingDetailState): MatchingDetailViewModel
+    }
+
+    companion object :
+        MavericksViewModelFactory<MatchingDetailViewModel, MatchingDetailState> by hiltMavericksViewModelFactory()
+}
+
