+++ feature/matching/src/main/java/com/puzzle/matching/page/detail/contract/MatchingDetailSideEffect.kt
@@ -1,0 +1,3 @@
+package com.puzzle.matching.page.detail.contract
+
+sealed class MatchingDetailSideEffect
