+++ feature/matching/src/main/java/com/puzzle/matching/page/main/contract/MatchingIntent.kt
@@ -1,0 +1,5 @@
+package com.puzzle.matching.page.main.contract
+
+sealed class MatchingIntent {
+    data object NavigateToMatchingDetail : MatchingIntent()
+}
