+++ feature/matching/src/main/java/com/puzzle/matching/page/main/contract/MatchingState.kt
@@ -1,0 +1,7 @@
+package com.puzzle.matching.page.main.contract
+
+import com.airbnb.mvrx.MavericksState
+
+data class MatchingState(
+    val isLoading: Boolean = false,
+) : MavericksState
