--- feature/matching/src/main/java/com/puzzle/matching/ui/DialogType.kt
@@ -1,7 +1,0 @@
-package com.puzzle.matching.ui
-
-enum class DialogType {
-    ACCEPT_MATCHING,
-    DECLINE_MATCHING,
-    PROFILE_IMAGE_DETAIL,
-}
