+++ feature/setting/src/main/java/com/puzzle/setting/navigation/SettingNavigation.kt
@@ -2,8 +2,8 @@ package com.puzzle.setting.navigation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.puzzle.setting.SettingRoute
 import com.puzzle.navigation.SettingRoute
+import com.puzzle.setting.SettingRoute
 
 fun NavGraphBuilder.settingScreen() {
     composable<SettingRoute> {
